services/k8s-copilot/requirements.txt
google-adk[a2a]
google-cloud-monitoring
google-cloud-logging
kubernetes
flask
gunicorn
pyyaml
google-auth
python-dotenv
